Be more aggressive when restarting fuseki
sometimes systemctl seems to miss that fuseki didn't shutdown

+++ cli/optimize.sh
@@ -4,7 +4,14 @@
 set -eu
 
 echo "Stopping Fuseki"
-systemctl stop fuseki
+FUSEKI_PID=$(systemctl show --property MainPID --help fuseki)
+FUSEKI_PID=${FUSEKI_PID#"MainPID="}
+if [ -n "${FUSEKI_PID}" ]; then
+    systemctl stop fuseki
+    # systemctl doesn't always seem to succeed, so be quite certain
+    # until we figure out why it doesn't always succeed
+    kill -9 "${FUSEKI_PID}" > /dev/null 2>&1
+fi
 
 RACK_DB="/etc/fuseki/databases/RACK"
 RACK_FUSEKI_CONFIG="/etc/fuseki/configuration/RACK.ttl"
